Reworked Vessel Entity Block and Recipe System
Overhauled how the Vessel works. Now it takes the Aspect from the items rather than the items themselves into it's inventory and increments or decreases the amount of Aspect it has based on Aspect given or taken away.

Also fixed rendering for the Aspects in Vessel for this new system and added a feature where you can see what Aspect a Foci currently has equipped. Foci can only have 1 Aspect at a time so it will remove any others if you try to give it more than 1 but I'm not sure how I feel about this, but it does make Spell Casting WAY easier this way so I'll probably keep it.